Cast RMSNorm weight to input dtype — enables F.rms_norm fused kernel
Without the cast, fp32 weight + bf16 input triggers a slow-path fallback,
defeating the optimization. The cast is on a small (D,) tensor per call.

+++ shared/components/norm.py
@@ -11,4 +11,7 @@ class RMSNorm(nn.Module):
         self.weight = nn.Parameter(torch.ones(dim))
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
-        return F.rms_norm(x, self.normalized_shape, self.weight, self.eps)
+        # Cast weight to input dtype so F.rms_norm can dispatch to fused kernel.
+        # Under bfloat16 autocast, x is bf16 but self.weight stays fp32; mismatch
+        # triggers a slow-path fallback otherwise. The cast is on a small (D,) tensor.
+        return F.rms_norm(x, self.normalized_shape, self.weight.to(x.dtype), self.eps)
